refactor #79 issue_constants.dart 위치 이동
